fix: 역 추가 후 재추가 안되는 버그 수정
