+++ src/managers/StationManager.js
@@ -1,8 +1,4 @@
-import {
-  rendStationMangeDom,
-  clearMangeContainer,
-  addStation,
-} from "../views/domController.js";
+import { clearMangeContainer } from "../views/domController.js";
 import { stationMangeContainer } from "../views/dom.js";
 import Station from "../components/Station.js";
 import { addLocalStorageByKey, deleteDataByName } from "../utils/util.js";
@@ -21,7 +17,7 @@ export default class StationManager {
     const stationNameElem = document.getElementById("station-name-input");
     if (addStationValidate(stationNameElem.value)) {
       addLocalStorageByKey("stations", new Station(stationNameElem.value));
-      rendStationMangeDom();
+      this.rendStationMangeDom();
     } else {
       alert(STATION.INPUT_ERROR_MESSAGE);
     }
@@ -32,8 +28,7 @@ export default class StationManager {
     try {
       if (deleteStationValidate(targetElem.dataset.index)) {
         deleteDataByName("stations", targetElem.dataset.index, "name");
-        const removeElem = targetElem.parentNode.parentNode;
-        removeElem.parentNode.removeChild(removeElem);
+        this.rendStationMangeDom();
       } else {
         alert(STATION.DELETE_ERROR_MESSAGE);
       }
@@ -44,7 +39,8 @@ export default class StationManager {
     document.querySelectorAll(".station-delete-button").forEach((item) => {
       item.addEventListener("click", (event) => {
         event.preventDefault();
-        if (confirm(DELETE_CONFIRM_MESSAGE)) confirmStationDelete(event.target);
+        if (confirm(DELETE_CONFIRM_MESSAGE))
+          this.confirmStationDelete(event.target);
       });
     });
   }
@@ -55,19 +51,20 @@ export default class StationManager {
     const div = document.createElement("div");
     div.innerHTML = stationMangeContainer();
     container.appendChild(div);
-    setStationDeleteEvent();
+    this.setStationDeleteEvent();
+    this.initEvent();
   }
 
   initEvent() {
     document
       .getElementById("station-add-button")
       .addEventListener("click", () => {
-        addStation();
+        this.addStation();
       });
   }
 
   render() {
-    rendStationMangeDom();
+    this.rendStationMangeDom();
     this.initEvent();
   }
 }
